requirements.txt
Django==1.10.1
djangorestframework==3.4.7
django-cors-headers==1.2.0
Pillow==3.3.1
qrcode==5.1
psycopg2==2.6.2